requirements.txt
asgiref==3.7.2
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
exceptiongroup==1.2.0
gunicorn==21.2.0
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
PyJWT==2.8.0
pytest==7.4.3
pytest-django==4.7.0
python-decouple==3.8
pytz==2023.3.post1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
